Address comments in PR #6.
Changes :
        - Move make_tls_obj() from https.HTTPSTransform to tlsutil for
          sharing
        - Fix 'is not' vs. '!=' in ftp.py
